requirements.txt
aiogram
apscheduler
python-dotenv
pytz
telethon
